+++ RV32_Cpu/Decoder/RV32D_Decoder.cs
@@ -13,7 +13,7 @@ namespace RiscVCpu.Decoder {
         /// <param name="instruction">32bit長の命令</param>
         /// <param name="cpu">命令を実行するRV32CPU</param>
         /// <returns>実行の成否</returns>
-        internal protected override bool Exec(byte[] ins, RV32_Cpu cpu) {
+        internal protected override bool Exec(UInt32[] ins, RV32_Cpu cpu) {
             bool result = false;
             FPRegister rd = (FPRegister)ins[1],
                         rs1 = (FPRegister)ins[3],
